+++ packages/amplify-graphql-api-construct/src/internal/index.ts
@@ -1,6 +1,5 @@
 export * from './authorization-modes';
 export * from './construct-exports';
-export * from './schema-preprocessors';
 export * from './default-parameters';
 export * from './conflict-resolution';
 export * from './asset-manager';