+++ packages/amplify-graphql-api-construct/src/types.ts
@@ -1,67 +1,118 @@
 import { Duration, CfnResource, NestedStack } from 'aws-cdk-lib';
 import {
-  SchemaFile,
   CfnGraphQLApi,
   CfnGraphQLSchema,
   CfnApiKey,
   CfnResolver,
   CfnFunctionConfiguration,
   CfnDataSource,
-  AppsyncFunctionProps,
   IGraphqlApi,
+  MappingTemplate,
 } from 'aws-cdk-lib/aws-appsync';
 import { CfnTable, ITable } from 'aws-cdk-lib/aws-dynamodb';
 import { IRole, CfnRole } from 'aws-cdk-lib/aws-iam';
 import { IUserPool } from 'aws-cdk-lib/aws-cognito';
 import { IFunction, CfnFunction } from 'aws-cdk-lib/aws-lambda';
 import { IBucket } from 'aws-cdk-lib/aws-s3';
-import { TransformerPluginProvider } from '@aws-amplify/graphql-transformer-interfaces';
-// These will be imported from CLI in future
-import { GraphqlOutput, BackendOutputStorageStrategy } from './graphql-output';
 
 /**
  * Configuration for IAM Authorization on the Graphql API.
+ * @struct - required since this interface begins with an 'I'
  */
-export type IAMAuthorizationConfig = {
-  identityPoolId?: string;
-  authenticatedUserRole?: IRole;
-  unauthenticatedUserRole?: IRole;
-  adminRoles?: IRole[];
-};
+export interface IAMAuthorizationConfig {
+  /**
+   * ID for the Cognito Identity Pool vending auth and unauth roles.
+   */
+  readonly identityPoolId?: string;
+
+  /**
+   * Authenticated user role, applies to { provider: iam, allow: private } access.
+   */
+  readonly authenticatedUserRole?: IRole;
+
+  /**
+   * Unauthenticated user role, applies to { provider: iam, allow: public } access.
+   */
+  readonly unauthenticatedUserRole?: IRole;
+
+  /**
+   * A list of roles granted full R/W access to the API.
+   */
+  readonly adminRoles?: IRole[];
+}
 
 /**
  * Configuration for Cognito UserPool Authorization on the Graphql API.
  */
-export type UserPoolAuthorizationConfig = {
-  userPool: IUserPool;
-};
+export interface UserPoolAuthorizationConfig {
+  /**
+   * The Cognito User Pool which is used to authenticated JWT tokens, and vends group and user information.
+   */
+  readonly userPool: IUserPool;
+}
 
 /**
  * Configuration for OpenId Connect Authorization on the Graphql API.
  */
-export type OIDCAuthorizationConfig = {
-  oidcProviderName: string;
-  oidcIssuerUrl: string;
-  clientId?: string;
-  tokenExpiryFromAuth: Duration;
-  tokenExpiryFromIssue: Duration;
-};
+export interface OIDCAuthorizationConfig {
+  /**
+   * The issuer for the OIDC configuration.
+   */
+  readonly oidcProviderName: string;
+
+  /**
+   * Url for the OIDC token issuer.
+   */
+  readonly oidcIssuerUrl: string;
+
+  /**
+   * The client identifier of the Relying party at the OpenID identity provider.
+   * A regular expression can be specified so AppSync can validate against multiple client identifiers at a time. Example
+   */
+  readonly clientId?: string;
+
+  /**
+   * The duration an OIDC token is valid after being authenticated by OIDC provider.
+   * auth_time claim in OIDC token is required for this validation to work.
+   */
+  readonly tokenExpiryFromAuth: Duration;
+
+  /**
+   * The duration an OIDC token is valid after being issued to a user.
+   * This validation uses iat claim of OIDC token.
+   */
+  readonly tokenExpiryFromIssue: Duration;
+}
 
 /**
  * Configuration for API Keys on the Graphql API.
  */
-export type ApiKeyAuthorizationConfig = {
-  description?: string;
-  expires: Duration;
-};
+export interface ApiKeyAuthorizationConfig {
+  /**
+   * Optional description for the api key to attach to the API.
+   */
+  readonly description?: string;
+
+  /**
+   * A duration representing the time from Cloudformation deploy until expiry.
+   */
+  readonly expires: Duration;
+}
 
 /**
  * Configuration for Custom Lambda authorization on the Graphql API.
  */
-export type LambdaAuthorizationConfig = {
-  function: IFunction;
-  ttl: Duration;
-};
+export interface LambdaAuthorizationConfig {
+  /**
+   * The authorizer lambda function.
+   */
+  readonly function: IFunction;
+
+  /**
+   * How long the results are cached.
+   */
+  readonly ttl: Duration;
+}
 
 /**
  * Authorization Config to apply to the API.
@@ -69,48 +120,42 @@ export type LambdaAuthorizationConfig = {
  * a defaultAuthMode must be specified.
  * For more information on Amplify API auth, refer to https://docs.amplify.aws/cli/graphql/authorization-rules/#authorization-strategies
  */
-export type AuthorizationConfig = {
+export interface AuthorizationConfig {
   /**
    * Default auth mode to provide to the API, required if more than one config type is specified.
    */
-  defaultAuthMode?: 'AWS_IAM' | 'AMAZON_COGNITO_USER_POOLS' | 'OPENID_CONNECT' | 'API_KEY' | 'AWS_LAMBDA';
+  readonly defaultAuthMode?: 'AWS_IAM' | 'AMAZON_COGNITO_USER_POOLS' | 'OPENID_CONNECT' | 'API_KEY' | 'AWS_LAMBDA';
 
   /**
    * IAM Auth config, required if an 'iam' auth provider is specified in the API.
    * Applies to 'public' and 'private' auth strategies.
    */
-  iamConfig?: IAMAuthorizationConfig;
+  readonly iamConfig?: IAMAuthorizationConfig;
 
   /**
    * Cognito UserPool config, required if a 'userPools' auth provider is specified in the API.
    * Applies to 'owner', 'private', and 'group' auth strategies.
    */
-  userPoolConfig?: UserPoolAuthorizationConfig;
+  readonly userPoolConfig?: UserPoolAuthorizationConfig;
 
   /**
    * Cognito OIDC config, required if a 'oidc' auth provider is specified in the API.
    * Applies to 'owner', 'private', and 'group' auth strategies.
    */
-  oidcConfig?: OIDCAuthorizationConfig;
+  readonly oidcConfig?: OIDCAuthorizationConfig;
 
   /**
    * AppSync API Key config, required if a 'apiKey' auth provider is specified in the API.
    * Applies to 'public' auth strategy.
    */
-  apiKeyConfig?: ApiKeyAuthorizationConfig;
+  readonly apiKeyConfig?: ApiKeyAuthorizationConfig;
 
   /**
    * Lambda config, required if a 'function' auth provider is specified in the API.
    * Applies to 'custom' auth strategy.
    */
-  lambdaConfig?: LambdaAuthorizationConfig;
-};
-
-/**
- * Conflict Handler Type for the DataSource
- * See https://docs.aws.amazon.com/appsync/latest/devguide/conflict-detection-and-sync.html#conflict-detection-and-resolution
- */
-export type ConflictHandlerType = 'OPTIMISTIC_CONCURRENCY' | 'AUTOMERGE' | 'LAMBDA';
+  readonly lambdaConfig?: LambdaAuthorizationConfig;
+}
 
 /**
  * Whether or not to use a version field to track conflict detection.
@@ -120,32 +165,51 @@ export type ConflictDetectionType = 'VERSION' | 'NONE';
 /**
  * Common parameters for conflict resolution.
  */
-export type ConflictResolutionStrategyBase = {
-  detectionType: ConflictDetectionType;
-  handlerType: ConflictHandlerType;
-};
+export interface ConflictResolutionStrategyBase {
+  /**
+   * The conflict detection type used for resolution.
+   */
+  readonly detectionType: ConflictDetectionType;
+}
 
 /**
  * Enable optimistic concurrency on the project.
  */
-export type AutomergeConflictResolutionStrategy = ConflictResolutionStrategyBase & {
-  handlerType: 'AUTOMERGE';
-};
+export interface AutomergeConflictResolutionStrategy extends ConflictResolutionStrategyBase {
+  /**
+   * This conflict resolution strategy executes an auto-merge.
+   * For more information, refer to https://docs.aws.amazon.com/appsync/latest/devguide/conflict-detection-and-sync.html#conflict-detection-and-resolution
+   */
+  readonly handlerType: 'AUTOMERGE';
+}
 
 /**
  * Enable automerge on the project.
  */
-export type OptimisticConflictResolutionStrategy = ConflictResolutionStrategyBase & {
-  handlerType: 'OPTIMISTIC_CONCURRENCY';
-};
+export interface OptimisticConflictResolutionStrategy extends ConflictResolutionStrategyBase {
+  /**
+   * This conflict resolution strategy the _version to perform optimistic concurrency.
+   * For more information, refer to https://docs.aws.amazon.com/appsync/latest/devguide/conflict-detection-and-sync.html#conflict-detection-and-resolution
+   */
+  readonly handlerType: 'OPTIMISTIC_CONCURRENCY';
+}
 
 /**
  * Enable custom sync on the project, powered by a lambda.
  */
-export type CustomConflictResolutionStrategy = ConflictResolutionStrategyBase & {
-  handlerType: 'LAMBDA';
-  conflictHandler: IFunction;
-};
+export interface CustomConflictResolutionStrategy extends ConflictResolutionStrategyBase {
+  /**
+   * This conflict resolution strategy uses a lambda handler type.
+   * For more information, refer to https://docs.aws.amazon.com/appsync/latest/devguide/conflict-detection-and-sync.html#conflict-detection-and-resolution
+
+   */
+  readonly handlerType: 'LAMBDA';
+
+  /**
+   * The function which will be invoked for conflict resolution.
+   */
+  readonly conflictHandler: IFunction;
+}
 
 /**
  * Conflict Resolution Strategy to apply to the project or a particular model.
@@ -158,82 +222,117 @@ export type ConflictResolutionStrategy =
 /**
  * Project level configuration for conflict resolution.
  */
-export type ConflictResolution = {
+export interface ConflictResolution {
   /**
    * Project-wide config for conflict resolution. Applies to all non-overridden models.
    */
-  project?: ConflictResolutionStrategy;
+  readonly project?: ConflictResolutionStrategy;
 
   /**
    * Model-specific conflict resolution overrides.
    */
-  models?: Record<string, ConflictResolutionStrategy>;
-};
-
-/**
- * Type representing Graphql representation of an amplify schema.
- */
-export type AmplifyApiGraphqlSchema = SchemaFile | SchemaFile[] | string;
+  readonly models?: Record<string, ConflictResolutionStrategy>;
+}
 
 /**
  * Custom type representing a processed schema output.
  */
-export type AmplifyApiSchemaPreprocessorOutput = {
+export interface AmplifyApiSchemaPreprocessorOutput {
   /**
    * Schema generated as an output of the preprocessing step.
    */
-  processedSchema: string;
+  readonly processedSchema: string;
 
   /**
    * Custom functions extracted during preprocessing.
    */
-  processedFunctionSlots?: FunctionSlot[];
-};
-
-/**
- * Custom preprocessor which will allow for custom schema representations that compile down to graphql
- * and custom resources (if necessary).
- */
-export type AmplifyApiSchemaPreprocessor<SchemaType> = (schema: SchemaType) => AmplifyApiSchemaPreprocessorOutput;
+  readonly processedFunctionSlots?: FunctionSlot[];
+}
 
 /**
  * Params exposed to support configuring and overriding pipelined slots. This allows configuration of the underlying function,
  * including the request and response mapping templates.
  */
-export type FunctionSlotOverride = Partial<Pick<AppsyncFunctionProps, 'requestMappingTemplate' | 'responseMappingTemplate'>>;
+export interface FunctionSlotOverride {
+  /**
+   * Override request mapping template for the function slot. Executed before the datasource is invoked.
+   */
+  readonly requestMappingTemplate?: MappingTemplate;
+
+  /**
+   * Override response mapping template for the function slot. Executed after the datasource is invoked.
+   */
+  readonly responseMappingTemplate?: MappingTemplate;
+}
 
 /**
  * Common slot parameters.
  */
-export type FunctionSlotBase = {
-  fieldName: string;
-  slotIndex: number;
-  function: FunctionSlotOverride;
-};
+export interface FunctionSlotBase {
+  /**
+   * The field to attach this function to on the schema.
+   */
+  readonly fieldName: string;
+
+  /**
+   * The slot index to use to inject this into the execution pipeline.
+   * For more information on slotting, refer to https://docs.amplify.aws/cli/graphql/custom-business-logic/#extend-amplify-generated-resolvers
+   */
+  readonly slotIndex: number;
+
+  /**
+   * The overridden behavior for this slot.
+   */
+  readonly function: FunctionSlotOverride;
+}
 
 /**
  * Slot types for Mutation Resolvers.
  */
-export type MutationFunctionSlot = FunctionSlotBase & {
-  typeName: 'Mutation';
-  slotName: 'init' | 'preAuth' | 'auth' | 'postAuth' | 'preUpdate' | 'postUpdate' | 'finish';
-};
+export interface MutationFunctionSlot extends FunctionSlotBase {
+  /**
+   * This slot type applies to the Mutation type on the schema.
+   */
+  readonly typeName: 'Mutation';
+
+  /**
+   * The slot name to inject this behavior into.
+   * For more information on slotting, refer to https://docs.amplify.aws/cli/graphql/custom-business-logic/#extend-amplify-generated-resolvers
+   */
+  readonly slotName: 'init' | 'preAuth' | 'auth' | 'postAuth' | 'preUpdate' | 'postUpdate' | 'finish';
+}
 
 /**
  * Slot types for Query Resolvers.
  */
-export type QueryFunctionSlot = FunctionSlotBase & {
-  typeName: 'Query';
-  slotName: 'init' | 'preAuth' | 'auth' | 'postAuth' | 'preDataLoad' | 'postDataLoad' | 'finish';
-};
+export interface QueryFunctionSlot extends FunctionSlotBase {
+  /**
+   * This slot type applies to the Query type on the schema.
+   */
+  readonly typeName: 'Query';
+
+  /**
+   * The slot name to inject this behavior into.
+   * For more information on slotting, refer to https://docs.amplify.aws/cli/graphql/custom-business-logic/#extend-amplify-generated-resolvers
+   */
+  readonly slotName: 'init' | 'preAuth' | 'auth' | 'postAuth' | 'preDataLoad' | 'postDataLoad' | 'finish';
+}
 
 /**
  * Slot types for Subscription Resolvers.
  */
-export type SubscriptionFunctionSlot = FunctionSlotBase & {
-  typeName: 'Subscription';
-  slotName: 'init' | 'preAuth' | 'auth' | 'postAuth' | 'preSubscribe';
-};
+export interface SubscriptionFunctionSlot extends FunctionSlotBase {
+  /**
+   * This slot type applies to the Subscription type on the schema.
+   */
+  readonly typeName: 'Subscription';
+
+  /**
+   * The slot name to inject this behavior into.
+   * For more information on slotting, refer to https://docs.amplify.aws/cli/graphql/custom-business-logic/#extend-amplify-generated-resolvers
+   */
+  readonly slotName: 'init' | 'preAuth' | 'auth' | 'postAuth' | 'preSubscribe';
+}
 
 /**
  * Input params to uniquely identify the slot which is being overridden.
@@ -246,55 +345,55 @@ export type FunctionSlot = MutationFunctionSlot | QueryFunctionSlot | Subscripti
  * a single location, and isn't spread around the transformers, where they can
  * have different default behaviors.
  */
-export type SchemaTranslationBehavior = {
+export interface SchemaTranslationBehavior {
   /**
    * Restore parity w/ GQLv1 @model parameter behavior, where setting a single field doesn't implicitly set the other fields to null.
    */
-  shouldDeepMergeDirectiveConfigDefaults: boolean;
+  readonly shouldDeepMergeDirectiveConfigDefaults: boolean;
 
   /**
    * Disable resolver deduping, this can sometimes cause problems because dedupe ordering isn't stable today, which can
    * lead to circular dependencies across stacks if models are reordered.
    */
-  disableResolverDeduping: boolean;
+  readonly disableResolverDeduping: boolean;
 
   /**
    * Enabling sandbox mode will enable api key auth on all models in the transformed schema.
    */
-  sandboxModeEnabled: boolean;
+  readonly sandboxModeEnabled: boolean;
 
   /**
    * Ensure that oidc and userPool auth use the `sub` field in the for the username field, which disallows new users with the same
    * id to access data from a deleted user in the pool.
    */
-  useSubUsernameForDefaultIdentityClaim: boolean;
+  readonly useSubUsernameForDefaultIdentityClaim: boolean;
 
   /**
    * Ensure that the owner field is still populated even if a static iam or group authorization applies.
    */
-  populateOwnerFieldForStaticGroupAuth: boolean;
+  readonly populateOwnerFieldForStaticGroupAuth: boolean;
 
   /**
    * If enabled, disable api key resource generation even if specified as an auth rule on the construct.
    * This is a legacy parameter from the Graphql Transformer existing in Amplify CLI, not recommended to change.
    */
-  suppressApiKeyGeneration: boolean;
+  readonly suppressApiKeyGeneration: boolean;
 
   /**
    * If disabled, generated @index as an LSI instead of a GSI.
    */
-  secondaryKeyAsGSI: boolean;
+  readonly secondaryKeyAsGSI: boolean;
 
   /**
    * Automate generation of query names, and as a result attaching all indexes as queries to the generated API.
    * If enabled, @index can be provided a null name field to disable the generation of the query on the api.
    */
-  enableAutoIndexQueryNames: boolean;
+  readonly enableAutoIndexQueryNames: boolean;
 
   /**
    * Enable custom primary key support, there's no good reason to disable this unless trying not to update a legacy app.
    */
-  respectPrimaryKeyAttributesOnConnectionField: boolean;
+  readonly respectPrimaryKeyAttributesOnConnectionField: boolean;
 
   /**
    * If enabled, set nodeToNodeEncryption on the searchable domain (if one exists). Not recommended for use, prefer
@@ -302,35 +401,143 @@ export type SchemaTranslationBehavior = {
    *   domain.NodeToNodeEncryptionOptions = { Enabled: True };
    * });
    */
-  enableSearchNodeToNodeEncryption: boolean;
-};
+  readonly enableSearchNodeToNodeEncryption: boolean;
+}
+
+/**
+ * A utility interface equivalent to Partial<SchemaTranslationBehavior>.
+ */
+export interface PartialSchemaTranslationBehavior {
+  /**
+   * Restore parity w/ GQLv1 @model parameter behavior, where setting a single field doesn't implicitly set the other fields to null.
+   */
+  readonly shouldDeepMergeDirectiveConfigDefaults?: boolean;
+
+  /**
+   * Disable resolver deduping, this can sometimes cause problems because dedupe ordering isn't stable today, which can
+   * lead to circular dependencies across stacks if models are reordered.
+   */
+  readonly disableResolverDeduping?: boolean;
+
+  /**
+   * Enabling sandbox mode will enable api key auth on all models in the transformed schema.
+   */
+  readonly sandboxModeEnabled?: boolean;
+
+  /**
+   * Ensure that oidc and userPool auth use the `sub` field in the for the username field, which disallows new users with the same
+   * id to access data from a deleted user in the pool.
+   */
+  readonly useSubUsernameForDefaultIdentityClaim?: boolean;
+
+  /**
+   * Ensure that the owner field is still populated even if a static iam or group authorization applies.
+   */
+  readonly populateOwnerFieldForStaticGroupAuth?: boolean;
+
+  /**
+   * If enabled, disable api key resource generation even if specified as an auth rule on the construct.
+   * This is a legacy parameter from the Graphql Transformer existing in Amplify CLI, not recommended to change.
+   */
+  readonly suppressApiKeyGeneration?: boolean;
+
+  /**
+   * If disabled, generated @index as an LSI instead of a GSI.
+   */
+  readonly secondaryKeyAsGSI?: boolean;
+
+  /**
+   * Automate generation of query names, and as a result attaching all indexes as queries to the generated API.
+   * If enabled, @index can be provided a null name field to disable the generation of the query on the api.
+   */
+  readonly enableAutoIndexQueryNames?: boolean;
+
+  /**
+   * Enable custom primary key support, there's no good reason to disable this unless trying not to update a legacy app.
+   */
+  readonly respectPrimaryKeyAttributesOnConnectionField?: boolean;
+
+  /**
+   * If enabled, set nodeToNodeEncryption on the searchable domain (if one exists). Not recommended for use, prefer
+   * to use `Object.values(resources.additionalResources['AWS::Elasticsearch::Domain']).forEach((domain: CfnDomain) => {
+   *   domain.NodeToNodeEncryptionOptions = { Enabled: True };
+   * });
+   */
+  readonly enableSearchNodeToNodeEncryption?: boolean;
+}
+
+/**
+ * Graphql schema definition, which can be implemented in multiple ways.
+ */
+export interface IAmplifyGraphqlSchema {
+  /**
+   * Return the schema definition as a graphql string.
+   * @returns the rendered schema.
+   */
+  readonly definition: string;
+
+  /**
+   * Retrieve any function slots defined explicitly in the schema.
+   * @returns generated function slots
+   */
+  readonly functionSlots: FunctionSlot[];
+}
+
+/**
+ * Entry representing the required output from the backend for codegen generate commands to work.
+ */
+export interface BackendOutputEntry {
+  /**
+   * The protocol version for this backend output.
+   */
+  readonly version: string;
+
+  /**
+   * The string-map payload of generated config values.
+   */
+  readonly payload: Record<string, string>;
+}
+
+/**
+ * Backend output strategy used to write config required for codegen tasks.
+ */
+export interface IBackendOutputStorageStrategy {
+  /**
+   * Add an entry to backend output.
+   * @param keyName the key
+   * @param strategy the backend output strategy information.
+   */
+  // eslint-disable-next-line @typescript-eslint/method-signature-style
+  addBackendOutputEntry(keyName: string, strategy: BackendOutputEntry): void;
+
+  /**
+   * Write all pending data to the destination
+   */
+  // eslint-disable-next-line @typescript-eslint/method-signature-style
+  flush(): void;
+}
 
 /**
  * Input props for the AmplifyGraphqlApi construct. Specifies what the input to transform into an API, and configurations for
  * the transformation process.
  */
-export type AmplifyGraphqlApiProps<SchemaType = AmplifyApiGraphqlSchema> = {
+export interface AmplifyGraphqlApiProps {
   /**
    * The schema to transform in a full API.
    */
-  schema: SchemaType;
-
-  /**
-   * Optional Preprocessor for the input schema, passes through a schema type to the
-   */
-  schemaPreprocessor?: AmplifyApiSchemaPreprocessor<SchemaType>;
+  readonly schema: IAmplifyGraphqlSchema;
 
   /**
    * Name to be used for the appsync api.
    * Default: construct id.
    */
-  apiName?: string;
+  readonly apiName?: string;
 
   /**
    * Required auth config for the API. This object must be a superset of the configured auth providers in the graphql schema.
    * For more information, refer to https://docs.amplify.aws/cli/graphql/authorization-rules/
    */
-  authorizationConfig: AuthorizationConfig;
+  readonly authorizationConfig: AuthorizationConfig;
 
   /**
    * Lambda functions referenced in the schema's @function directives. The keys of this object are expected to be the
@@ -338,13 +545,13 @@ export type AmplifyGraphqlApiProps<SchemaType = AmplifyApiGraphqlSchema> = {
    * map, then it is interpreted as the `functionName`, and an arn will be constructed using the current aws account and region
    * (or overridden values, if set in the directive).
    */
-  functionNameMap?: Record<string, IFunction>;
+  readonly functionNameMap?: Record<string, IFunction>;
 
   /**
    * Configure conflict resolution on the API, which is required to enable DataStore API functionality.
    * For more information, refer to https://docs.amplify.aws/lib/datastore/getting-started/q/platform/js/
    */
-  conflictResolution?: ConflictResolution;
+  readonly conflictResolution?: ConflictResolution;
 
   /**
    * StackMappings override the assigned nested stack on a per-resource basis. Only applies to resolvers, and takes the form
@@ -353,124 +560,124 @@ export type AmplifyGraphqlApiProps<SchemaType = AmplifyApiGraphqlSchema> = {
    * after initial deployment AppSync resolvers cannot be moved between nested stacks, they will need to be removed from the app,
    * then re-added from a new stack.
    */
-  stackMappings?: Record<string, string>;
+  readonly stackMappings?: Record<string, string>;
 
   /**
    * Overrides for a given slot in the generated resolver pipelines. For more information about what slots are available,
    * refer to https://docs.amplify.aws/cli/graphql/custom-business-logic/#override-amplify-generated-resolvers.
    */
-  functionSlots?: FunctionSlot[];
+  readonly functionSlots?: FunctionSlot[];
 
   /**
    * Provide a list of additional custom transformers which are injected into the transform process.
    */
-  transformers?: TransformerPluginProvider[];
+  readonly transformers?: any[];
 
   /**
    * If using predictions, a bucket must be provided which will be used to search for assets.
    */
-  predictionsBucket?: IBucket;
+  readonly predictionsBucket?: IBucket;
 
   /**
    * This replaces feature flags from the API construct, for general information on what these parameters do,
    * refer to https://docs.amplify.aws/cli/reference/feature-flags/#graphQLTransformer
    */
-  schemaTranslationBehavior?: Partial<SchemaTranslationBehavior>;
+  readonly schemaTranslationBehavior?: PartialSchemaTranslationBehavior;
 
   /**
    * Strategy to store construct outputs. If no outputStorageStrategey is provided a default strategy will be used.
    */
-  outputStorageStrategy?: BackendOutputStorageStrategy<GraphqlOutput>;
-};
+  readonly outputStorageStrategy?: IBackendOutputStorageStrategy;
+}
 
 /**
  * L1 CDK resources from the API which were generated as part of the transform.
  * These are potentially stored under nested stacks, but presented organized by type instead.
  */
-export type AmplifyGraphqlApiCfnResources = {
+export interface AmplifyGraphqlApiCfnResources {
   /**
    * The Generated AppSync API L1 Resource
    */
-  cfnGraphqlApi: CfnGraphQLApi;
+  readonly cfnGraphqlApi: CfnGraphQLApi;
 
   /**
    * The Generated AppSync Schema L1 Resource
    */
-  cfnGraphqlSchema: CfnGraphQLSchema;
+  readonly cfnGraphqlSchema: CfnGraphQLSchema;
 
   /**
    * The Generated AppSync API Key L1 Resource
    */
-  cfnApiKey?: CfnApiKey;
+  readonly cfnApiKey?: CfnApiKey;
 
   /**
    * The Generated AppSync Resolver L1 Resources, keyed by logicalId.
    */
-  cfnResolvers: Record<string, CfnResolver>;
+  readonly cfnResolvers: Record<string, CfnResolver>;
 
   /**
    * The Generated AppSync Function L1 Resources, keyed by logicalId.
    */
-  cfnFunctionConfigurations: Record<string, CfnFunctionConfiguration>;
+  readonly cfnFunctionConfigurations: Record<string, CfnFunctionConfiguration>;
 
   /**
    * The Generated AppSync DataSource L1 Resources, keyed by logicalId.
    */
-  cfnDataSources: Record<string, CfnDataSource>;
+  readonly cfnDataSources: Record<string, CfnDataSource>;
 
   /**
    * The Generated DynamoDB Table L1 Resources, keyed by logicalId.
    */
-  cfnTables: Record<string, CfnTable>;
+  readonly cfnTables: Record<string, CfnTable>;
 
   /**
    * The Generated IAM Role L1 Resources, keyed by logicalId.
    */
-  cfnRoles: Record<string, CfnRole>;
+  readonly cfnRoles: Record<string, CfnRole>;
 
   /**
    * The Generated Lambda Function L1 Resources, keyed by function name.
    */
-  cfnFunctions: Record<string, CfnFunction>;
+  readonly cfnFunctions: Record<string, CfnFunction>;
 
   /**
    * Remaining L1 resources generated, keyed by logicalId.
    */
-  additionalCfnResources: Record<string, CfnResource>;
-};
+  readonly additionalCfnResources: Record<string, CfnResource>;
+}
 
 /**
  * Accessible resources from the API which were generated as part of the transform.
  * These are potentially stored under nested stacks, but presented organized by type instead.
  */
-export type AmplifyGraphqlApiResources = {
+export interface AmplifyGraphqlApiResources {
   /**
    * The Generated AppSync API L2 Resource, includes the Schema.
    */
-  graphqlApi: IGraphqlApi;
+  readonly graphqlApi: IGraphqlApi;
 
   /**
    * The Generated DynamoDB Table L2 Resources, keyed by logicalId.
    */
-  tables: Record<string, ITable>;
+  readonly tables: Record<string, ITable>;
 
   /**
    * The Generated IAM Role L2 Resources, keyed by logicalId.
    */
-  roles: Record<string, IRole>;
+  readonly roles: Record<string, IRole>;
 
   /**
    * The Generated Lambda Function L1 Resources, keyed by function name.
    */
-  functions: Record<string, IFunction>;
+  readonly functions: Record<string, IFunction>;
 
   /**
    * L1 Cfn Resources, for when dipping down a level of abstraction is desirable.
    */
-  cfnResources: AmplifyGraphqlApiCfnResources;
+  readonly cfnResources: AmplifyGraphqlApiCfnResources;
 
   /**
    * Nested Stacks generated by the API Construct.
    */
-  nestedStacks: Record<string, NestedStack>;
-};
+  readonly nestedStacks: Record<string, NestedStack>;
+}