+++ packages/amplify-category-api/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [5.6.4](https://github.com/aws-amplify/amplify-category-api/compare/@aws-amplify/amplify-category-api@5.6.3...@aws-amplify/amplify-category-api@5.6.4) (2023-09-13)
+
+**Note:** Version bump only for package @aws-amplify/amplify-category-api
+
+## [5.6.3](https://github.com/aws-amplify/amplify-category-api/compare/@aws-amplify/amplify-category-api@5.6.2...@aws-amplify/amplify-category-api@5.6.3) (2023-09-07)
+
+**Note:** Version bump only for package @aws-amplify/amplify-category-api
+
 ## [5.6.2](https://github.com/aws-amplify/amplify-category-api/compare/@aws-amplify/amplify-category-api@5.6.1...@aws-amplify/amplify-category-api@5.6.2) (2023-08-30)
 
 **Note:** Version bump only for package @aws-amplify/amplify-category-api